refactor: move util to path_utils
