+++ src/griptape_nodes/files/path_utils.py
@@ -15,9 +15,31 @@ and are used by OSManager, FileDrivers, and workspace managers.
 import os
 import re
 import sys
-from pathlib import Path
+from pathlib import Path, PurePosixPath
 from typing import NamedTuple
 from urllib.parse import unquote, urlparse
+
+
+def derive_registry_key(file_path: str) -> str:
+    """Derive a workflow registry key from a file path.
+
+    Strips the file extension and normalizes directory separators to forward slashes,
+    preserving directory components for uniqueness across different directories.
+
+    Args:
+        file_path: Path to the workflow file, e.g. "subdir/my_workflow.py"
+
+    Returns:
+        Registry key with directory components preserved, e.g. "subdir/my_workflow"
+
+    Examples:
+        >>> derive_registry_key("my_workflow.py")
+        "my_workflow"
+        >>> derive_registry_key("subdir/my_workflow.py")
+        "subdir/my_workflow"
+    """
+    normalized = file_path.replace("\\", "/")
+    return str(PurePosixPath(normalized).with_suffix(""))
 
 
 class FilenameParts(NamedTuple):
